backend/requirements.txt
# ===============================
# Core API
# ===============================
fastapi[standard]>=0.128.0
uvicorn[standard]>=0.30.3
python-multipart>=0.0.9
httpx>=0.27.2

# Database
# ===============================
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
asyncpg>=0.30.0
alembic>=1.13.3

# ===============================
# Auth & Security
# ===============================
python-jose[cryptography]>=3.3.0
passlib[argon2,bcrypt]>=1.7.4
argon2-cffi>=25.1.0
bcrypt>=4.2.0
cryptography>=43.0.1

# ===============================
# Pydantic (PYTHON 3.13 SAFE)
# ===============================
pydantic>=2.9.2
pydantic-settings>=2.6.0
email-validator>=2.2.0
python-dotenv>=1.0.1

# ===============================
# AI / LLM
# ===============================
google-generativeai>=0.7.2
sentence-transformers>=3.1.1
langchain>=0.2.14
langchain-pinecone>=0.2.13
langchain-text-splitters>=1.1.0

# ===============================
# Document Processing
# ===============================
PyPDF2>=3.0.1
python-docx>=1.1.2
pdfplumber>=0.11.4
pillow>=12.0.0
python-magic>=0.4.27
pymupdf>=1.26.7

# ===============================
# AWS / Storage
# ===============================
boto3>=1.34.160

# ===============================
# Background Tasks & Cache
# ===============================
# redis>=5.0.8
# celery>=5.4.0

# ===============================
# Monitoring & Logging
# ===============================
prometheus-client>=0.21.0
sentry-sdk>=2.8.0
structlog>=24.4.0

# ===============================
# Rate Limiting
# ===============================
slowapi>=0.1.9
limits>=3.12.0

# ===============================
# Testing & Dev Tools
# ===============================
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
black>=24.10.0
isort>=5.13.2
flake8>=7.1.1
mypy>=1.13.0
pre-commit>=4.0.1


